chore(*) refactor helm/helm into helm/
This resolves issue #90.
