+++ manifest/manifest_test.go
