+++ testdata/charts/dep3/Chart.yaml
