+++ testdata/charts/kitchensink/Chart.yaml
