+++ testdata/charts/kitchensink/annoy_testers/broken.yaml
