+++ testdata/charts/kitchensink/manifests/foo.json
