+++ testdata/charts/kitchensink/manifests/nested/nested-pod.yaml
