+++ testdata/charts/kitchensink/manifests/sink-daemonset.yaml-wait-till-1.1
