+++ testdata/charts/kitchensink/manifests/sink-namespace.yaml
