+++ testdata/charts/kitchensink/manifests/sink-pod.yaml
