+++ testdata/charts/kitchensink/manifests/sink-rc.yaml
