+++ testdata/charts/kitchensink/manifests/sink-secret.yaml
