+++ testdata/charts/kitchensink/manifests/sink-server.yaml
