+++ testdata/charts/kitchensink/manifests/sink-volume.yaml
