+++ testdata/charts/redis/Chart.yaml
