+++ testdata/charts/redis/manifests/redis-pod.yaml
