+++ testdata/charts/searchtest1/manifests/redis-pod.yaml
