+++ testdata/charts/searchtest2/manifests/redis-pod.yaml
